feat: reviewinfo dto 별점들 실수형으로 변경

+++ app/src/main/java/com/eatssu/android/data/model/response/GetReviewInfoResponseDto.kt
@@ -3,9 +3,9 @@ package com.eatssu.android.data.model.response
 data class GetReviewInfoResponseDto(
     val menuName: List<String>,
     val totalReviewCount: Int,
-    val mainGrade: Int,
-    val amountGrade: Int,
-    val tasteGrade: Int,
+    val mainGrade: Double,
+    val amountGrade: Double,
+    val tasteGrade: Double,
     val reviewGradeCnt: ReviewGradeCnt,
 ){
     data class ReviewGradeCnt(
